% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ax.R
\name{normalise_sex}
\alias{normalise_sex}
\title{Normalise a sex argument to 1 (male) or 2 (female)}
\usage{
normalise_sex(sex)
}
\arguments{
\item{sex}{\code{1}, \code{2}, \code{"male"}, \code{"female"}, \code{"m"} or \code{"f"}.}
}
\value{
\code{1L} or \code{2L}.
}
\description{
Normalise a sex argument to 1 (male) or 2 (female)
}
